Fixed new compile-time errors for Xcode 10.3 beta.

+++ framework/Source/CPTGradient.m
@@ -1002,7 +1002,7 @@ static void CPTResolveHSV(CGFloat *__nonnull color1, CGFloat *__nonnull color2);
         return (NSUInteger)theHash;
     }
     else {
-        return self.blendingMode + self.gradientType;
+        return (NSUInteger)(self.blendingMode + self.gradientType);
     }
 }
